requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# LLM API
# Google Gemini client
google-generativeai>=0.7.2

# Data Processing
pandas==2.1.1
numpy==1.24.3
openpyxl==3.1.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Scientific Computing
scipy==1.11.3
scikit-learn==1.3.0

# Environment Management
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Security
cryptography==41.0.4

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1